+++ kitty/window.py
@@ -117,6 +117,7 @@ if TYPE_CHECKING:
 
     from .fast_data_types import MousePosition
     from .file_transmission import FileTransmission
+    from .notifications import OnlyWhen
 
 
 class CwdRequestType(Enum):
@@ -630,7 +631,7 @@ class Window:
         self.current_mouse_event_button = 0
         self.current_clipboard_read_ask: Optional[bool] = None
         self.last_cmd_output_start_time = 0.
-        self.last_notification_id: Optional[int] = None
+        self.last_cmd_end_notification: Optional[tuple[int, 'OnlyWhen']] = None
         self.open_url_handler: 'OpenUrlHandler' = None
         self.last_cmd_cmdline = ''
         self.last_cmd_exit_status = 0
@@ -1214,13 +1215,12 @@ class Window:
                 tab = self.tabref()
                 if tab is not None:
                     tab.relayout_borders()
-            if self.last_notification_id:
-                # Notification id is saved withing handle_cmd_end so it
-                # configured to be close upon focus is gained and visibility
-                # change. When window is focused, it is visible for sure.
-                nm = get_boss().notification_manager
-                nm.close_notification(self.last_notification_id)
-                self.last_notification_id = None
+            if self.last_cmd_end_notification is not None:
+                from .notifications import OnlyWhen
+                opts = get_options()
+                if self.last_cmd_end_notification[1] in (OnlyWhen.unfocused, OnlyWhen.invisible) and 'focus' in opts.notify_on_cmd_finish.clear_on:
+                    get_boss().notification_manager.close_notification(self.last_cmd_end_notification[0])
+                    self.last_cmd_end_notification = None
         elif self.os_window_id == current_focused_os_window_id():
             # Cancel IME composition after loses focus
             update_ime_position_for_window(self.id, False, -1)
@@ -1518,7 +1518,7 @@ class Window:
             "is_start": False, "time": end_time, 'cmdline': self.last_cmd_cmdline, 'exit_status': self.last_cmd_exit_status})
 
         opts = get_options()
-        when, duration, action, notify_cmdline = opts.notify_on_cmd_finish
+        when, duration, action, notify_cmdline, _ = opts.notify_on_cmd_finish
 
         if last_cmd_output_duration >= duration and when != 'never':
             from .notifications import OnlyWhen
@@ -1531,20 +1531,13 @@ class Window:
             if not nm.is_notification_allowed(cmd, self.id):
                 return
             if action == 'notify':
-                # Notification id is saved, so configuration was checked on
-                # previous pass and saved to be cleared upon focus. But that
-                # action was missed somehow and we should clear it here.
-                if self.last_notification_id:
-                    nm.close_notification(self.last_notification_id)
-                    self.last_notification_id = None
+                if self.last_cmd_end_notification is not None:
+                    if 'next' in opts.notify_on_cmd_finish.clear_on:
+                        nm.close_notification(self.last_cmd_end_notification[0])
+                    self.last_cmd_end_notification = None
                 notification_id = nm.notify_with_command(cmd, self.id)
-                # Saving notification id only when we are going to close it in
-                # future.
-                # TODO(Shvedov): We should close notification not only when
-                # gather focus, but when window become visible if `when` equals
-                # to "invisible".
-                if cmd.only_when is OnlyWhen.unfocused or cmd.only_when is OnlyWhen.invisible:
-                    self.last_notification_id = notification_id
+                if notification_id is not None:
+                    self.last_cmd_end_notification = notification_id, cmd.only_when
             elif action == 'bell':
                 self.screen.bell()
             elif action == 'command':